refactor: Bio 컴포넌트 개선 - user prop 제거
Phase 3 완료: Bio 컴포넌트에서 SITE_CONFIG 직접 사용

## 변경 사항

### Bio 컴포넌트
- user prop 제거
- 내부에서 SITE_CONFIG.author 직접 사용
- User 객체 변환 로직 내부화
- 'use client' 유지 (클라이언트 컴포넌트)

### AboutPage 컴포넌트
- user 객체 생성 로직 제거
- Bio 컴포넌트 호출 간소화: <Bio size="lg" />
- User import 제거

## 효과
✅ Props 간소화 (user prop 제거)
✅ 데이터 접근 통일 (SITE_CONFIG 직접 사용)
✅ AboutPage 코드 간결화
✅ 빌드 성공

## 완료된 작업
- Phase 1-5: Props Drilling 제거 ✅
- Phase 3: Bio 컴포넌트 개선 ✅
