+++ src/page-components/about/ui/AboutPage.tsx
@@ -1,5 +1,4 @@
 import { Bio } from "@/entities/user";
-import { User } from "@/entities/user/model/types";
 import { Thread, ThreadCard } from "@/entities/thread";
 import { Badge } from "@/shared/ui/badge";
 import { Card, CardContent } from "@/shared/ui/card";
@@ -15,29 +14,13 @@ export function AboutPage({
   threads = [],
   className,
 }: AboutPageProps) {
-  const user: User = {
-    id: SITE_CONFIG.author.id,
-    name: SITE_CONFIG.author.name,
-    email: SITE_CONFIG.author.email,
-    avatar: SITE_CONFIG.author.avatar,
-    bio: {
-      role: SITE_CONFIG.author.bio.role,
-      description: [...SITE_CONFIG.author.bio.description],
-      location: SITE_CONFIG.author.bio.location,
-    },
-    social: {
-      github: SITE_CONFIG.author.socialLinks.github,
-      linkedin: SITE_CONFIG.author.socialLinks.linkedin,
-    },
-  };
-
   return (
     <Layout className={className}>
       <div className="container mx-auto px-4 py-8 max-w-2xl">
         <section className="mb-8">
           <Card>
             <CardContent className="pt-6">
-              <Bio user={user} size="lg" />
+              <Bio size="lg" />
 
               {SITE_CONFIG.author.skills.length > 0 && (
                 <div className="flex flex-wrap gap-2 mt-6 justify-center">
